+++ ididi/_itypes.py
@@ -8,6 +8,7 @@ from typing import (
     Generator,
     Protocol,
     TypedDict,
+    TypeVar,
     Union,
     overload,
     runtime_checkable,
@@ -19,7 +20,6 @@ EMPTY_SIGNATURE = inspect.Signature()
 INSPECT_EMPTY = inspect.Signature.empty
 
 IEmptyFactory = Callable[[], R]
-IEmptyAsyncFactory = Callable[[], Awaitable[R]]
 
 IFactory = Callable[P, R]
 IAnyFactory = Callable[..., R]
@@ -42,8 +42,20 @@ INodeFactory = Union[
 ]
 INode = Union[INodeFactory[P, R], type[R]]
 
+P1 = TypeVar("P1")
+P2 = TypeVar("P2")
+P3 = TypeVar("P3")
+P4 = TypeVar("P4")
+P5 = TypeVar("P5")
+P6 = TypeVar("P6")
+P7 = TypeVar("P7")
+P8 = TypeVar("P8")
+P9 = TypeVar("P9")
 
+
+# Factory with many type params
 class TDecor(Protocol):
+
     @overload
     def __call__(self, factory: type[T]) -> type[T]: ...
 
@@ -63,6 +75,10 @@ class TDecor(Protocol):
 
     @overload
     def __call__(self, factory: INode[P, T]) -> INode[P, T]: ...
+
+
+class TEntryDecor(Protocol):
+    def __call__(self, func: Callable[P, T]) -> Callable[..., T]: ...
 
 
 class INodeConfig(TypedDict, total=False):