+++ ididi/api.py
@@ -1,12 +1,28 @@
-from typing import Any, Awaitable, Callable, Union, overload
+from functools import partial
+from typing import Any, Awaitable, Callable, Union, cast, overload
 
+from typing_extensions import Unpack
+
+from ._itypes import IAsyncFactory, IEmptyFactory, IFactory, INodeConfig, TEntryDecor
 from .graph import DependencyGraph as DependencyGraph
 from .utils.typing_utils import P, T
 
 
-def entry(func: Callable[P, T]) -> Union[Callable[..., T], Callable[..., Awaitable[T]]]:
-    dg = DependencyGraph()
-    return dg.entry(func)
+@overload
+def entry(**iconfig: Unpack[INodeConfig]) -> TEntryDecor: ...
+
+
+@overload
+def entry(func: IFactory[P, T]) -> IEmptyFactory[T]: ...
+
+
+def entry(
+    func: Union[IFactory[P, T], IAsyncFactory[P, T], None] = None,
+    **iconfig: Unpack[INodeConfig],
+) -> Union[Callable[..., Union[T, Awaitable[T]]], TEntryDecor]:
+    if not func:
+        return cast(TEntryDecor, partial(entry, **iconfig))
+    return DependencyGraph().entry(func, **iconfig)
 
 
 @overload
@@ -18,5 +34,4 @@ def resolve(dep: Callable[P, T], /, *args: P.args, **overrides: P.kwargs) -> T: 
 
 
 def resolve(dep: Callable[P, T], /, **overrides: Any) -> T:
-    dg = DependencyGraph()
-    return dg.resolve(dep, **overrides)
+    return DependencyGraph().resolve(dep, **overrides)