+++ tests/features/test_inject.py
@@ -1,16 +1,18 @@
-# from ididi import DependencyGraph, entry, inject
-# from tests.features.services import UserService
+from ididi import DependencyGraph, entry, inject
+from tests.test_data import UserService
 
 
-# def get_user_service() -> UserService:
-#     dg = DependencyGraph()
-#     return dg.resolve(UserService)
+def get_user_service(dg: DependencyGraph) -> UserService:
+    return UserService(1, 2)
 
 
-# async def create_user(service: UserService = inject(get_user_service)):
-#     assert isinstance(service, UserService)
+async def create_user(service: UserService = inject(get_user_service)):
+    assert isinstance(service, UserService)
+    assert service.db == 1
+    assert service.auth == 2
+    return "ok"
 
 
-# async def test_inject_entry():
-#     f = entry(create_user)
-#     await f()
+async def test_inject_entry():
+    f = entry(create_user)
+    assert await f() == "ok"